Finished "NoPad" flag to only round up to nearest even number
FFT routine takes only even-sized Fourier vector, so it is
impossible to have no padding if the number of time steps is odd.
However, we now had a flag "NoPad" that only evens out the
number, it does not feature any other padding. The original
flag "ALL" still also works.