Created temporal coherence module
Separate file TemporalCoherence.f90 contains the routine to set
the Rho and Mu arrays based on the temporal coherence flag in the
input. The grid parameters are also fed in so can specify the
values spatially later. Phase sampling routine is also changed
now to call this function and use the Rho/Mu values during the
phase sampling. Quickly verified using python routine.